Use relative coordinates for game objects
I made two major, related changes. First, I've unified map and
game object (e.g. Obstacle, Player, GameLocation) coordinates.
Previously, maps were rendered on a scale of 32 pixels (that is,
a map location 1,1 was rendered at 32,32) while object locations
were specified directly in pixels 532,1029 -> 532,1029. Now,
object coordinates are a continuous version of map coordinates,
e.g. object location 1.02,3.2 -> [x-scale]*1.02,[y-scale]*3.2.

Second, the rendering unit is now dependent on the screen size.
Previously, everything was rendered on a 32,32 unit scale, which
worked great for my computer and badly for anyone with a different
desktop configuration. The scale is now 1) determined by the map
dimensions and screen size 2) determined separately for the
vertical and horizontal dimensions. This results in distortion for
some screens, but, short of creating multiple map versions, I don't
see a way around that.

To facilitate these changes, I brought back the Renderer, which is
more or less a thin wrapper around GraphicsContext that is aware
of the rendering coordinates.

The majority of these changes are either some form of dividing
coordinate values by 32, or integerating with Renderer.