Let player movement depend on time, not game ticks
While testing the game on my desktop, I found that that the game
framerate appears to vary inversely with the window size, that is,
it would run 3x slower with a larger screen. The effect could be
duplicated by resizing the game window.

This happened JavaFX AnimationTimer framerate cannot be fixed. It is
controlled only by some internal clock, which seems to slow down
with large display sizes.

To fix this, I'm just taking the time elapsed since the previous
frame into acount when moving players.

An unintended result is that player movement is visibly smoother
now.